fix(zmq): bump the wave clock on every lockstep wake
## Description

### Problem

The dp lane hung on gpt-oss with `no engine output for 300s`, and the
engine's own log shows why: `EngineCore_DP0` died on
`TimeoutError: RPC call to sample_tokens timed out` after five minutes of
`No available shared memory broadcast block found in 60 seconds`. Rank 0
was blocked in the model forward waiting for rank 1 to join the
all-reduce; rank 1 was parked and never woke.

The wake was dropped on the floor. `START_DP_WAVE` is ignored by an
engine whose `current_wave` already exceeds the wave it names
(`core.py:_handle_client_request`), and the ranks bump `current_wave`
themselves as they park (`core.py:run_busy_loop`). So between the group
draining a wave and this client processing the resulting
`wave_complete`, the client's clock reads one wave behind the engines,
and the next submit's wake is stale — silently discarded. The holder
still gets its Add, starts stepping, and blocks forever on peers that
were never told to wake. Waking unconditionally (d71f4af3) did not help:
the message was sent, just ignored.

### Solution

Treat the wave number as a logical clock this client owns, and bump it on
every wake rather than replaying the last number the engines reported.
That is provably never stale: a rank sits at the last number we broadcast
plus at most one self-increment (a second increment needs another wave,
which needs another wake), so `last + 1` always clears it. Reported
waves are still folded in, upward only, so a rank found running ahead
(kept its `current_wave` across a gateway restart) also gets cleared.

The wake now goes to every rank, with no excluded index. Excluding the
request holder left it on a lower `current_wave` than its peers, which
breaks the argument above for the price of one saved message.

`WaveState` collapses to the bare clock: its `running` flag existed for
the wake gate removed in d71f4af3. The clock widens to u64 so the
monotonic bump cannot wrap a long-lived gateway into the same lost-wake
hang.

## Changes

- `connector.rs`: `wake_group` bumps the clock and broadcasts to all
  ranks; `observe_wave` folds reported waves upward only; `WaveState` is
  now a `Mutex<u64>`.
- `protocol/mod.rs`, `vllm/mod.rs`, `tokenspeed/mod.rs`:
  `encode_start_wave(wave: u64)` drops the exclusion parameter and sends
  a sentinel index that matches no rank.
- `vllm/output.rs`, `mock_engine.rs`: wave fields widened to u64.

## Test Plan

- `cargo test -p engine-zmq-client` — 72 passing.
  `every_wake_names_a_wave_the_ranks_will_accept` replaces
  `a_drained_wave_re_arms_the_wake`: it asserts each wake strictly
  outranks the previous one without waiting for any `wave_complete`, and
  that a rank reporting a wave ahead of the clock is cleared by the next
  wake.
- `cargo clippy --workspace --all-targets -- -D warnings`.
- `e2e-2gpu-chat-zmq-dp (vllm)` is the live exercise — this is the lane
  that hung.

## Checklist

- [x] Tests added/updated
- [x] Lint and formatting pass
- [x] No breaking changes to public APIs

+++ crates/engine_zmq_client/src/protocol/vllm/output.rs
@@ -143,18 +143,18 @@ struct WireEngineCoreOutputs {
     finished_requests: Option<BTreeSet<String>>,
     /// In DP mode, signals that the current wave finished and engines are paused.
     #[serde(default)]
-    wave_complete: Option<u32>,
+    wave_complete: Option<u64>,
     /// In DP mode, signals that a request arrived for an old wave and the next
     /// wave needs to start in other engines.
     #[serde(default)]
-    start_wave: Option<u32>,
+    start_wave: Option<u64>,
 }
 
 /// Data-parallel control notifications multiplexed through `EngineCoreOutputs`.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum DpControlMessage {
-    WaveComplete(u32),
-    StartWave(u32),
+    WaveComplete(u64),
+    StartWave(u64),
 }
 
 /// A batch of per-request outputs plus the piggybacked scheduler stats.
